Add Game component
The Game component houses the game's core logic, including tracking
clicked cards, managing win conditions, updating scores, and triggering
game resets. It is comprised of a Scoreboard and a grid of Cards.

Data for the Pokemon cards is fetched asynchronously from the PokéAPI.
The useEffect hook is employed to restrict data refreshes to the
beginning of each game round. Additionally, an error message is
displayed if data retrieval fails.